requirements.txt
absl-py==2.1.0
ale-py==0.8.1
bindsnet @ git+https://github.com/BindsNET/bindsnet.git@1b9786152db6fc67a38d2ddb3090ac434bca239f
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpickle==3.1.0
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
distro==1.9.0
eagerpy==0.30.0
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.54.1
foolbox==3.3.4
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.67.1
gymnasium==0.29.1
idna==3.10
imageio==2.36.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pyaml==24.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-build==0.18.1
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
Shimmy==0.2.1
six==1.16.0
smmap==5.0.1
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2024.9.20
tomli==2.0.2
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.6
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.2